Fix harmless compiler warnings.
FossilOrigin-Name: 9a93c68a7673bda2e6c308d3b2798d0fb12fde399996cae24b1f0d5864adf355

+++ ext/fts3/fts3_aux.c
@@ -417,14 +417,14 @@ static int fts3auxFilterMethod(
     if( zStr ){
       pCsr->filter.zTerm = sqlite3_mprintf("%s", zStr);
       if( pCsr->filter.zTerm==0 ) return SQLITE_NOMEM;
-      pCsr->filter.nTerm = strlen(pCsr->filter.zTerm);
+      pCsr->filter.nTerm = (int)strlen(pCsr->filter.zTerm);
     }
   }
 
   if( iLe>=0 ){
     pCsr->zStop = sqlite3_mprintf("%s", sqlite3_value_text(apVal[iLe]));
     if( pCsr->zStop==0 ) return SQLITE_NOMEM;
-    pCsr->nStop = strlen(pCsr->zStop);
+    pCsr->nStop = (int)strlen(pCsr->zStop);
   }
   
   if( iLangid>=0 ){
